Add environment variable mapping (#1)
By default Pi-Hole is intended to be exposed behind a transparent proxy gateway such as nginx. In case such a proxy is not available for some reason, set compose environment variable WEB_MAP to a port mapping to 80, e.g. "WEB_MAP=- '8080:80'".

Note that the port mapping value must be quoted !